Move Plans from openapi spec to crd
