+++ threescaleAPI-operator/pkg/apis/3scale/v1alpha1/types.go
@@ -2,6 +2,7 @@ package v1alpha1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sort"
 )
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
@@ -31,7 +32,36 @@ type APISpec struct {
 	The3ScaleConfig   The3ScaleConfig `json:"3scaleConfig"`
 	Upstream          string          `json:"upstream"`
 	OpenAPIDefinition string          `json:"OpenAPIDefinition"`
+	Plans             Plans           `json:"plans"`
 }
 type APIStatus struct {
 	// Fill me
+}
+
+// TODO - If Plans are expected to be common across backends, this should be moved from here as appropriate
+type Plans []Plan
+
+func (p Plans) Sort() Plans {
+	for _, plan := range p {
+		sort.Slice(plan.Limits, func(i, j int) bool {
+			if plan.Limits[i].Metric != plan.Limits[j].Metric {
+				return plan.Limits[i].Metric < plan.Limits[j].Metric
+			} else {
+				return plan.Limits[i].Max < plan.Limits[j].Max
+			}
+		})
+	}
+	return p
+}
+
+type Plan struct {
+	Default bool    `json:"default, omitempty"`
+	Name    string  `json:"name"`
+	Limits  []Limit `json:"limits"`
+}
+
+type Limit struct {
+	Max    int64  `json:"max"`
+	Metric string `json:"metric"`
+	Period string `json:"period"`
 }
