+++ threescaleAPI-operator/pkg/threescaleAPI/reconciliation.go
@@ -5,6 +5,7 @@ import (
 	"encoding/xml"
 	"errors"
 	"fmt"
+	"github.com/3scale/ostia/threescaleAPI-operator/pkg/apis/3scale/v1alpha1"
 	"github.com/3scale/ostia/threescaleAPI-operator/pkg/threescale/system_client"
 	"github.com/getkin/kin-openapi/jsoninfo"
 	"github.com/getkin/kin-openapi/openapi3"
@@ -45,7 +46,7 @@ func ensureServiceExists(c *client.ThreeScaleClient, accessToken string, name st
 	return svc, nil
 }
 
-func reconcilePlansAndLimits(c *client.ThreeScaleClient, service client.Service, accessToken string, desiredPlans Plans) {
+func reconcilePlansAndLimits(c *client.ThreeScaleClient, service client.Service, accessToken string, desiredPlans v1alpha1.Plans) {
 	var planFetchErr error
 
 	type mappingRuleStore map[string]struct {
@@ -89,7 +90,7 @@ func reconcilePlansAndLimits(c *client.ThreeScaleClient, service client.Service,
 	go func() {
 		for k, v := range havePlans {
 			keep := false
-			for _, plan := range desiredPlans.Plans {
+			for _, plan := range desiredPlans {
 				if plan.Name == k {
 					keep = true
 					break
@@ -107,7 +108,7 @@ func reconcilePlansAndLimits(c *client.ThreeScaleClient, service client.Service,
 	//mappingRuleValues := <-mrStoreC
 
 	// for all the plans we want
-	for _, wantedPlan := range desiredPlans.Plans {
+	for _, wantedPlan := range desiredPlans {
 		var id string
 		// Have you read this plan name previously? If its in the map then yes, set the id
 		if val, ok := havePlans[wantedPlan.Name]; ok {
@@ -239,22 +240,6 @@ func getAppPlans(c *client.ThreeScaleClient, svcId, accessToken string) (map[str
 	return nameToIds, nil
 
 }
-func decodePlans(s *openapi3.Swagger) (Plans, error) {
-	var desiredPlans Plans
-
-	switch s.Extensions["x-3scale-plans"].(type) {
-	case json.RawMessage:
-		err := json.Unmarshal(s.Extensions["x-3scale-plans"].(json.RawMessage), &desiredPlans)
-		if err != nil {
-			return desiredPlans, errors.New("error calling unmarshal of plans")
-
-		}
-	default:
-		return desiredPlans, errors.New("error - plans not recognised as json")
-
-	}
-	return desiredPlans, nil
-}
 
 func reconcileEndpointsWith3scaleSystem(c *client.ThreeScaleClient, accessToken string, service client.Service, existingEndpoints Endpoints, desiredEndpoints Endpoints) error {
 
@@ -308,7 +293,7 @@ func reconcileEndpointsWith3scaleSystem(c *client.ThreeScaleClient, accessToken 
 
 }
 
-func comparePlans(planA Plans, planB Plans) bool {
+func comparePlans(planA v1alpha1.Plans, planB v1alpha1.Plans) bool {
 
 	A, _ := json.Marshal(planA.Sort())
 	B, _ := json.Marshal(planB.Sort())
@@ -469,22 +454,22 @@ func getEndpointsFromSwagger(swagger *openapi3.Swagger) (Endpoints, error) {
 
 	return endpoints, nil
 }
-func getPlansFrom3scaleSystem(c *client.ThreeScaleClient, accessToken string, service client.Service) (Plans, error) {
+func getPlansFrom3scaleSystem(c *client.ThreeScaleClient, accessToken string, service client.Service) (v1alpha1.Plans, error) {
 
-	var plans Plans
+	var plans v1alpha1.Plans
 
 	appPlans, _ := c.ListAppPlanByServiceId(accessToken, service.ID)
 
 	for _, v := range appPlans.Plans {
 
-		var plan Plan
+		var plan v1alpha1.Plan
 
 		limits, _ := c.ListLimitsPerAppPlan(accessToken, v.ID)
 
 		plan.Name = v.PlanName
 		metrics, _ := c.ListMetrics(accessToken, v.ServiceID)
 		for _, v := range limits.Limits {
-			var limit Limit
+			var limit v1alpha1.Limit
 			metricID := v.MetricID
 			var metricName string
 			for _, v := range metrics.Metrics {
@@ -499,21 +484,10 @@ func getPlansFrom3scaleSystem(c *client.ThreeScaleClient, accessToken string, se
 			plan.Limits = append(plan.Limits, limit)
 
 		}
-		plans.Plans = append(plans.Plans, plan)
+		plans = append(plans, plan)
 	}
 
 	return plans, nil
-}
-func getPlansFromSwagger(swagger *openapi3.Swagger) (Plans, error) {
-	var desiredPlans Plans
-
-	s, _ := json.Marshal(swagger.Extensions["x-3scale-plans"])
-	err := json.Unmarshal(s, &desiredPlans)
-	if err != nil {
-		return desiredPlans, err
-	}
-
-	return desiredPlans, nil
 }
 
 // getNonDesiredMetricsFromSystem returns a list of metrics that are in the current system but not desired
